+++ Changelog.md
@@ -1,3 +1,9 @@
+## Qlibs 0.6.0
+ - Feature: SpriteMaster can now load PIL images and ImageData directly.
+ - Feature: texture filter can now be specified for SpriteMaster.
+ - Feature: absolute paths can be used instead of resource paths everywhere.
+ - Fix: multiple fonts not working. At all.
+
 ## Qlibs 0.5.7
  - Fix: turns out glyph.bearing.x was ignored. Not anymore, resulting in a more correct font rendering.